backend/requirements.txt
fastapi[standard]
python-jose[cryptography]
bcrypt